% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_data.R
\docType{data}
\name{data_quadas}
\alias{data_quadas}
\title{Example QUADAS-2 assessment dataset}
\format{
A data frame with the following 6 columns:
\describe{
  \item{Study}{Study identifier}
  \item{D1}{Domain 1 - Patient Selection}
  \item{D2}{Domain 2 - Index Test}
  \item{D3}{Domain 3 - Reference Standard}
  \item{D4}{Domain 4 -  Flow & Timing}
  \item{Overall}{Overall risk of bias}
         }
}
\source{
{Created for this package}
}
\usage{
data_quadas
}
\description{
A data frame containing 12 example assessments performed using
the risk-of-bias portion of the QUADAS-2 tool for the assessment of
diagnostic accuracy studies.
}
\keyword{datasets}
