% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_data.R
\docType{data}
\name{data_rob2_cluster}
\alias{data_rob2_cluster}
\title{Example ROB2.0 (cluster variant) assessment dataset}
\format{
A data frame with the following 7 columns:

\describe{
  \item{Study}{Study identifier}
  \item{D1}{Domain 1 - Bias arising from the randomization process}
  \item{D1b}{Domain 1b - Bias arising from the timing of identification and recruitment of Individual participants in relation to timing of randomization}
  \item{D2}{Domain 2 - Bias due to deviations from intended intervention}
  \item{D3}{Domain 3 - Bias due to missing outcome data}
  \item{D4}{Domain 4 - Bias in measurement of the outcome}
  \item{D5}{Domain 5 - Bias in selection of the reported result}
  \item{Overall}{Overall risk of bias}
         }
}
\source{
{Created for this package}
}
\usage{
data_rob2_cluster
}
\description{
A data frame containing 9 example assessments performed using
the cluster-randomised version of the RoB 2.0 tool.
}
\keyword{datasets}
