% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_robins.R
\docType{data}
\name{data_robins}
\alias{data_robins}
\title{Example ROBINS-I assessment}
\format{
A data frame:
\describe{
  \item{Study}{Study details}
  \item{D1}{Domain 1}
  \item{D2}{Domain 2}
  \item{D3}{Domain 3}
  \item{D4}{Domain 4}
  \item{D5}{Domain 5}
  \item{D6}{Domain 6}
  \item{D7}{Domain 7}
  \item{Overall}{Overall risk of bias}
  \item{Weight}{Weight measure for each study}
         }
}
\source{
{Created for this package}
}
\usage{
data_robins
}
\description{
Example ROBINS-I assessment
}
\keyword{internal}
