% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_data.R
\docType{data}
\name{data_robins_i}
\alias{data_robins_i}
\title{Example ROBINS-I assessment dataset}
\format{
A data frame with the following 9 columns:

\describe{
  \item{Study}{Study identifier}
  \item{D1}{Domain 1 - Bias due to confounding}
  \item{D2}{Domain 2 - Bias due to selection of participants}
  \item{D3}{Domain 3 - Bias in classification of interventions}
  \item{D4}{Domain 4 - Bias due to deviations from intended interventions}
  \item{D5}{Domain 5 - Bias due to missing data}
  \item{D6}{Domain 6 - Bias in measurement of outcomes}
  \item{D7}{Domain 7 - Bias in selection of the reported result}
  \item{Overall}{Overall risk of bias}
         }
}
\source{
{Created for this package}
}
\usage{
data_robins_i
}
\description{
A data frame containing 10 example assessments performed using
the ROBINS-I (Risk Of Bias In Non-randomised Studies - of Interventions) tool.
}
\keyword{datasets}
