% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_append_to_forest.R
\name{rob_append_to_forest}
\alias{rob_append_to_forest}
\title{Append a risk-of-bias traffic-light plot to a forest plot}
\usage{
rob_append_to_forest(
  res,
  rob_tool = "ROB2",
  rob_colour = "cochrane",
  rob_psize = 2,
  rob_caption = FALSE,
  rob_legend = TRUE,
  legend_cex = 0.9,
  ...
)
}
\arguments{
\item{res}{Output from metafor meta-analysis function}

\item{rob_tool}{The risk-of-bias assessment tool used to perform the
assessments}

\item{rob_colour}{Risk of bias colour scheme to use for the traffic-light
plot}

\item{rob_psize}{Size of the points in the appended risk of bias table}

\item{rob_caption}{Logical specifying whether a message containing the
risk-of-bias domains should be printed. Default is TRUE}

\item{rob_legend}{Logical specifying whether a legend for the risk-of-bias
plot should be shown. Default is TRUE.}

\item{...}{Additional arguments to be passed to the metafor::forest()
function}

\item{rob_data}{Risk of bias dataset in standard robvis format}
}
\description{
A wrapper for metafor::forest function, which adds a risk of
  bias traffic-light plot to the right-hand side of the forest plot. The
  heavy lifting for this function is done by metafor. Note that if not
  specified as additional arguments, this functions sets the header argument
  of metafor::forest() to TRUE.
}
\seealso{
Other main: 
\code{\link{rob_summary}()},
\code{\link{rob_traffic_light}()}
}
\concept{main}
