% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_tools.R
\name{rob_append_weights}
\alias{rob_append_weights}
\title{Extract weights from metafor results object and append to risk-of-bias data.}
\usage{
rob_append_weights(data, res)
}
\arguments{
\item{data}{Risk of bias dataset (without a weight column)}

\item{res}{metafor results object}
}
\description{
Used to prepare a risk-of-bias dataset to be passed to the
  weighted barplot function: rob_summary(..., weighted = TRUE)
}
\examples{
\donttest{
dat.bcg <- metadat::dat.bcg[c(1:9),]

dat <-
  metafor::escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg,
    slab = paste(author, year)
  )

res <- metafor::rma(yi, vi, data = dat)

data_rob2$Study <- paste(dat$author,dat$year)

rob_weighted_data <- rob_append_weights(data_rob2[,1:7], res)

rob_summary(rob_weighted_data, tool = "ROB2", weighted = TRUE)
}
}
\seealso{
Other helper:
\code{\link[=rob_dummy]{rob_dummy()}},
\code{\link[=rob_save]{rob_save()}},
\code{\link[=rob_tools]{rob_tools()}}
}
\concept{helper}
