% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_direction.R
\name{rob_direction}
\alias{rob_direction}
\title{Bias direction plots}
\usage{
rob_direction(
  dat,
  vi = NULL,
  sei = NULL,
  title = NULL,
  legend_cex = 0.9,
  grouping = "type",
  grouping_levels = c("MR", "NRSI", "NRSE", "RCT"),
  label_subgroup_summary = "RE Model for Subgroup",
  ...
)
}
\arguments{
\item{dat}{Dataframe}

\item{vi}{Vector containing the sampling variances (normally defined as the column within the dataset, i.e. dat$vi). Note: either vi or sei must be set.}

\item{sei}{Vector containing the corresponding standard errors (normally defined as the column within the dataset, i.e. dat$sei). Note: either vi or sei must be set.}

\item{title}{Graph title}

\item{legend_cex}{Expansion factor for figure legend.}

\item{grouping}{Variable of the provided dataset by which the resulting plot will be stratified. Often will study design or overall risk-of-bias level.}

\item{grouping_levels}{Ordering of grouping variable. Note: the levels will be plotted
in order, starting at the bottom of the graph (i.e. the last item in the
vector will be placed at the top of the graph)}

\item{label_subgroup_summary}{Annotation text for subgroup label}

\item{...}{Other arguments to pass to metafor::forest}
}
\description{
Used to
}
