% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_paired_plot.R
\name{rob_paired_plot}
\alias{rob_paired_plot}
\title{Produce forest plots with risk of bias results}
\usage{
rob_paired_plot(data, tool, colour = "cochrane", psize = 20,
  quiet = FALSE)
}
\arguments{
\item{data}{A dataframe containing summary (domain) level risk-of-bias assessments, with the first column containing the study details, the second column containing the first domain of your assessments, and the final column containing a weight to assign to each study. The function assumes that the data includes a column for overall risk-of-bias. For example, a ROB2.0 dataset would have 8 columns (1 for study details, 5 for domain level judgements, 1 for overall judgements, and 1 for weights, in that order).}
}
\description{
A function to take a summary table of risk of bias assessments and produce a traffic light plot from it.
}
