% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_summary_tools.R
\name{rob_summary_tools}
\alias{rob_summary_tools}
\title{Produce summary figure of risk of bias assessments.}
\usage{
rob_summary_tools()
}
\value{
Risk of bias assessment figure in .png output.
}
\description{
A function to convert standard risk-of-bias output to tidy data .
}
