% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robvis.R
\docType{package}
\name{robvis}
\alias{robvis}
\alias{robvis-package}
\title{robvis: A package for producing risk of bias assessment summary figures.}
\description{
The robvis package provides two useful functions to produce summary level
risk of bias assessment figures.
}
\section{robvis functions}{

The functions ...
}

