% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quadas.R
\docType{data}
\name{data_quadas}
\alias{data_quadas}
\title{Example QUADAS-2 assessment}
\format{
An object of class \code{data.frame} with 12 rows and 7 columns.
}
\source{
{Created for this package}
}
\usage{
data_quadas
}
\description{
#' @format A data frame:
\describe{
  \item{Study}{Study details}
  \item{D1}{Domain 1}
  \item{D2}{Domain 2}
  \item{D3}{Domain 3}
  \item{D4}{Domain 4}
  \item{Overall}{Overall risk of bias}
  \item{Weight}{Weight measure for each study}
         }
}
\keyword{datasets}
