% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_data.R
\docType{data}
\name{data_quips}
\alias{data_quips}
\title{Example QUIPS assessment}
\format{
A data frame:
\describe{
  \item{Study}{Study details}
  \item{D1}{Domain 1}
  \item{D2}{Domain 2}
  \item{D3}{Domain 3}
  \item{D4}{Domain 4}
  \item{D5}{Domain 5}
  \item{D6}{Domain 6}
  \item{Overall}{Overall risk of bias}
         }
}
\source{
{Created for this package}
}
\usage{
data_quips
}
\description{
Example QUIPS assessment
}
\keyword{internal}
