% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rob1.R
\docType{data}
\name{data_rob1}
\alias{data_rob1}
\title{Example ROB1 assessment}
\format{
An object of class \code{data.frame} with 9 rows and 10 columns.
}
\source{
{Created for this package}
}
\usage{
data_rob1
}
\description{
#' @format A data frame:
\describe{
  \item{Study}{Study details}
  \item{D1}{Domain 1}
  \item{D2}{Domain 2}
  \item{D3}{Domain 3}
  \item{D4}{Domain 4}
  \item{D5}{Domain 5}
  \item{D6}{Domain 6}
  \item{D7}{Domain 7}
  \item{Overall}{Overall risk of bias}
  \item{Weight}{Weight measure for each study}
         }
}
\keyword{datasets}
