% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rob1.R
\docType{data}
\name{data_rob1}
\alias{data_rob1}
\title{Example ROB1 assessment}
\format{
A data frame:
\describe{
  \item{Study}{Study details}
  \item{Random.sequence.generation.}{Domain 1}
  \item{Allocation.concealment.}{Domain 2}
  \item{Blinding.of.participants.and.personnel.}{Domain 3}
  \item{Blinding.of.outcome.assessment}{Domain 4}
  \item{Incomplete.outcome.data}{Domain 5}
  \item{Selective.reporting.}{Domain 6}
  \item{Other.sources.of.bias.}{Domain 7}
  \item{Overall}{Overall risk of bias}
  \item{Weight}{Weight measure for each study}
         }
}
\source{
{Created for this package}
}
\usage{
data_rob1
}
\description{
Example ROB1 assessment
}
\keyword{internal}
