% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rob2.R
\docType{data}
\name{data_rob2}
\alias{data_rob2}
\title{Example ROB2.0 assessment}
\format{
A data frame:
\describe{
  \item{Study}{Study details}
  \item{D1}{Domain 1}
  \item{D2}{Domain 2}
  \item{D3}{Domain 3}
  \item{D4}{Domain 4}
  \item{D5}{Domain 5}
  \item{Overall}{Overall risk of bias}
  \item{Weight}{Weight measure for each study}
         }
}
\source{
{Created for this package}
}
\usage{
data_rob2
}
\description{
Example ROB2.0 assessment
}
\keyword{internal}
