% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_data.R
\docType{data}
\name{data_rob2_cluster}
\alias{data_rob2_cluster}
\title{Example ROB2.0 assessment}
\format{
A data frame:
\describe{
  \item{Study}{Study details}
  \item{D1}{Domain 1}
  \item{D1b}{Domain 1b}
  \item{D2}{Domain 2}
  \item{D3}{Domain 3}
  \item{D4}{Domain 4}
  \item{D5}{Domain 5}
  \item{Overall}{Overall risk of bias}
         }
}
\source{
{Created for this package}
}
\usage{
data_rob2_cluster
}
\description{
Example ROB2.0 assessment
}
\keyword{internal}
