% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_blobbogram.R
\name{rob_blobbogram}
\alias{rob_blobbogram}
\title{Risk of bias blobbogram plot}
\usage{
rob_blobbogram(
  rma,
  rob,
  rob_tool = "ROB2",
  rob_colour = "cochrane",
  subset_col = "Overall",
  space_last = TRUE,
  subset_col_order = NULL,
  overall_estimate = TRUE,
  add_tests = TRUE,
  ...
)
}
\arguments{
\item{rma}{an rma object from the metafor package}

\item{rob}{a risk of bias table, either ROB2, ROB2-Cluster, QUADAS-2, or Robins (see example data)}

\item{rob_tool}{The risk of bias tool used. Defaults to ROB2. Supports "ROB2", "ROB2-Cluster", "QUADAS-2", and "Robins".}

\item{rob_colour}{The risk of bias colour scheme. As in other robvis functions. Supports "cochrane", "colourblind" or a custom vector of colours.}

\item{subset_col}{Column containing the variable to subset the meta-analysis by. Default "Overall".}

\item{subset_col_order}{An optional vector containing the levels of "subset_col" in the order that they should appear in the table.}

\item{overall_estimate}{Logical. Show a summary measure across groups?}

\item{add_tests}{Logical. Should a summary of statistical tests by subgroup be included as a row?}

\item{...}{options to be passed to forester::forester. See https://github.com/rdboyes/forester for more information.}
}
\value{
an image
}
\description{
Risk of bias blobbogram plot
}
