% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_blobbogram.R
\name{rob_blobbogram}
\alias{rob_blobbogram}
\title{Risk of bias blobbogram plot}
\usage{
rob_blobbogram(
  rma,
  rob,
  rob_tool = "ROB2",
  rob_colour = "cochrane",
  subset_col = "Overall",
  add_tests = TRUE,
  overall_estimate = TRUE,
  theme = NULL,
  ggplot_is_x_times_right_width = 1.2,
  null_line_at = 1,
  file_path = tempfile(pattern = "rob_blobbogram", fileext = ".png"),
  dpi = 600,
  display = TRUE,
  blank_na = TRUE,
  font_family = "mono",
  estimate_col_name = "Estimate"
)
}
\arguments{
\item{rma}{an rma object from the metafor package}

\item{rob}{a risk of bias table formatted like data_rob2}

\item{rob_tool}{risk of bias tool used}

\item{rob_colour}{risk of bias colour scheme}

\item{subset_col}{Column containing the variable to subset the MA by}

\item{add_tests}{include summary of statistical tests by subgroup}

\item{overall_estimate}{show a summary measure across groups?}

\item{theme}{optional theme for the forest plot table}

\item{ggplot_is_x_times_right_width}{relative width of the forest plot}

\item{null_line_at}{what value is the null? default 1.0}

\item{file_path}{path to save the image}

\item{dpi}{image resolution}

\item{display}{should the image be shown in the viewer? set to false for markdown}

\item{blank_na}{show NA values? default is FALSE}

\item{font_family}{(experimental) change fonts in the table}

\item{estimate_col_name}{title of the estimate column, default "Estimate"}
}
\value{
an image
}
\description{
Risk of bias blobbogram plot
}
