% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_dummy.R
\name{rob_dummy}
\alias{rob_dummy}
\title{Create "realistic" dummy risk of bias assessment data}
\usage{
rob_dummy(n, tool)
}
\arguments{
\item{n}{Number of assessments to create}

\item{tool}{Tool used for assessment (see rob_tools())}
}
\description{
Create "realistic" dummy risk of bias assessment data
}
