% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_forest.R
\name{rob_forest}
\alias{rob_forest}
\title{Append a risk-of-bias traffic-light plot to a forest plot}
\usage{
rob_forest(
  res,
  rob_tool = "ROB2",
  rob_me = NULL,
  rob_levels = NULL,
  title = NULL,
  rob_legend = TRUE,
  rob_legend_cex = 0.9,
  ...
)
}
\arguments{
\item{res}{Output from metafor meta-analysis function}

\item{rob_tool}{The risk-of-bias assessment tool used to perform the
assessments}

\item{rob_levels}{Vector of judgments [e.g. c("Low","Some
concerns","High","Critical")] that controls the ordering of subgroups
within the plot}

\item{title}{Text to use for plot title}

\item{rob_legend}{Logical specifying whether a legend for the risk-of-bias
plot should be shown. Default is TRUE.}

\item{rob_legend_cex}{Expansion factor for the risk-of-bias legend}

\item{...}{Additional arguments to be passed to the metafor::forest()
function}
}
\description{
A wrapper for metafor::forest function, which adds a risk of
  bias traffic-light plot to the right-hand side of the forest plot. The
  heavy lifting for this function is done by metafor. Note that if not
  specified as additional arguments, this functions sets the header argument
  of metafor::forest() to TRUE.
}
\seealso{
Other main: 
\code{\link{rob_summary}()},
\code{\link{rob_traffic_light}()}
}
\concept{main}
