% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_save.R
\name{rob_save}
\alias{rob_save}
\title{Save risk-of-bias plots to a file using sensible parameters}
\usage{
rob_save(
  rob_object,
  file = "rob_figure.png",
  height = "default",
  width = "default",
  dpi = 800
)
}
\arguments{
\item{rob_object}{Object created using either rob_summary() or
rob_traffic_light()}

\item{file}{Destination file, with extension (e.g. "rob_figure.png")}

\item{height}{Height of resulting image, in inches. Defaults to "default"
which uses recommended values based on the number of studies included.}

\item{width}{Width of resulting image, in inches. Defaults to "default" which
uses recommended values based on the number of characters in the Study and
Domain names. dpi}

\item{dpi}{Plot resolution.}
}
\description{
Save risk-of-bias plots to a file using sensible parameters
}
\seealso{
Other helper: 
\code{\link{rob_append_weights}()},
\code{\link{rob_dummy}()},
\code{\link{rob_tools}()}
}
\concept{helper}
