% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_summary_plot.R
\name{rob_summary}
\alias{rob_summary}
\title{Produce summary figure of risk of bias assessments.}
\usage{
rob_summary(data, tool, save = "No")
}
\arguments{
\item{data}{A .csv file of your summary (domain) level risk of bias assessments, with the first column containing the study name, and the second column containing the first domain of your assessments.}

\item{tool}{The risk of bias assessment tool used. RoB2.0 (tool="ROB2"), ROBINS-I (tool="ROBINS-I"), and QUADAS-2 (tool="QUADAS-2") are currently supported.}

\item{save}{An option to save the plot as the specified file type. Default is "No", and available extensions are eps/ps, tex (pictex), pdf, jpeg, tiff, png, bmp, svg and wmf (windows only).}
}
\value{
Risk of bias assessment figure in .png output.
}
\description{
TBC.
}
