% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_summary_tools.R
\name{rob_summary_tools}
\alias{rob_summary_tools}
\title{List tools covered by rob_summary().}
\usage{
rob_summary_tools()
}
\description{
rob_sumary_tools() will list the tools that can currently be plotted using the rob_summary() function.
}
