% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_tools.R
\name{rob_tools}
\alias{rob_tools}
\title{List tools covered by rob_summary().}
\usage{
rob_tools()
}
\description{
rob_tools() will list the tools that can currently be plotted using the rob_summary() function.
}
\examples{
rob_tools()
}
