% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_tools.R
\name{rob_tools}
\alias{rob_tools}
\title{List tool templates available within the package.}
\usage{
rob_tools(forest = FALSE)
}
\arguments{
\item{forest}{Show tools supported by the forest/blobbogram functions}
}
\description{
rob_tools() will list the risk-of-bias assessment tools for
  which templates already exist within the robvis package. If the assessment
  tool you used does not appear in the list, use the "Generic" template.
}
\examples{
rob_tools()
}
\seealso{
Other helper: 
\code{\link{rob_save}()}
}
\concept{helper}
