% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_traffic_light.R
\name{rob_traffic_light}
\alias{rob_traffic_light}
\title{Produce traffic-light plots of risk-of-bias assessments.}
\usage{
rob_traffic_light(data, tool, quiet = FALSE)
}
\arguments{
\item{data}{A dataframe containing summary (domain) level risk-of-bias assessments, with the first column containing the study details, the second column containing the first domain of your assessments, and the final column containing a weight to assign to each study. The function assumes that the data includes a column for overall risk-of-bias. For example, a ROB2.0 dataset would have 8 columns (1 for study details, 5 for domain level judgements, 1 for overall judgements, and 1 for weights, in that order).}

\item{tool}{The risk of bias assessment tool used. RoB2.0 (tool="ROB2"), ROBINS-I (tool="ROBINS-I"), and QUADAS-2 (tool="QUADAS-2") are currently supported.}

\item{quiet}{An option to quietly produce and save the plot without it displaying in R/Rstudio.}
}
\value{
Risk-of-bias assessment traffic light plot ggplot2::ggplot2 object
}
\description{
A function to take a summary table of risk of bias assessments and produce a traffic light plot from it.
}
