% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_traffic_light.R
\name{rob_traffic_light}
\alias{rob_traffic_light}
\title{Produce traffic-light plots of risk-of-bias assessments.}
\usage{
rob_traffic_light(
  data,
  tool,
  colour = "cochrane",
  psize = 20,
  quiet = FALSE,
  x_title = "Risk of bias domains",
  y_title = "Study",
  judgement_labels = c("Critical", "High", "Unclear", "Low", "No information",
    "Not applicable"),
  judgement_title = "Judgement"
)
}
\arguments{
\item{data}{A dataframe containing summary (domain) level risk-of-bias assessments, with the first column containing the study details, the second column containing the first domain of your assessments, and the final column containing a weight to assign to each study. The function assumes that the data includes a column for overall risk-of-bias. For example, a ROB2.0 dataset would have 8 columns (1 for study details, 5 for domain level judgments, 1 for overall judgements, and 1 for weights, in that order).}

\item{tool}{The risk of bias assessment tool used. RoB2.0 (tool='ROB2'), ROBINS-I (tool='ROBINS-I'), and QUADAS-2 (tool='QUADAS-2') are currently supported.}

\item{colour}{An argument to specify the colour scheme for the plot. Default is 'cochrane' which used the ubiquitous Cochrane colours, while a preset option for a colour-blind friendly palette is also available (colour = 'colourblind').}

\item{psize}{Control the size of the traffic lights. Default is 20.}

\item{quiet}{An option to quietly produce the plot without displaying it.}

\item{x_title}{Optional argument to alter x axis title, only applicable if using generic template}

\item{y_title}{Optional argument to alter y axis title, only applicable if using generic template}

\item{judgement_labels}{Optional argument to customise judgement labels, should be in order of c(critical, high, unclear, low, no information, not applicable), only applicable if using generic template.}

\item{judgement_title}{Optional argument to customise the title above judgement labels, only applicable if using generic template}
}
\value{
Risk-of-bias assessment traffic light plot (ggplot2 object)
}
\description{
A function to take a summary table of risk of bias assessments and produce a traffic light plot from it.
}
\examples{

data <- data.frame(stringsAsFactors=FALSE,
                   Study = c("Study 1", "Study 2"),
                   D1 = c("Low", "Some concerns"),
                   D2 = c("Low", "Low"),
                   D3 = c("Low", "Low"),
                   D4 = c("Low", "Low"),
                   D5 = c("Low", "Low"),
                   Overall = c("Low", "Low"),
                   Weight = c(33.33333333, 33.33333333)
                   )

rob_traffic_light(data, "ROB2")

}
