% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robvis.R
\docType{package}
\name{robvis-package}
\alias{robvis}
\alias{robvis-package}
\title{robvis: A package for producing risk-of-bias assessment figures.}
\description{
The robvis package is designed to help users produce publication quality
risk-of-bias assessment figures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcguinlu/robvis}
  \item Report bugs at \url{https://github.com/mcguinlu/robvis}
}

}
\author{
\strong{Maintainer}: Luke McGuinness \email{luke.mcguinness@bristol.ac.uk}

Authors:
\itemize{
  \item Luke McGuinness \email{luke.mcguinness@bristol.ac.uk}
}

Other contributors:
\itemize{
  \item Emily Kothe [contributor]
  \item Alex Fowler [contributor]
  \item Randall Boyes [contributor]
  \item Wolfgang Viechtbauer [contributor]
}

}
\keyword{internal}
