% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robvis.R
\docType{package}
\name{robvis}
\alias{robvis}
\alias{robvis-package}
\title{robvis: A package for producing risk-of-bias assessment figures.}
\description{
The robvis package is designed to help users produce publication quality
risk-of-bias assessment figures.
}
